+++ projects/rocprofiler-compute/CHANGELOG.md
@@ -115,6 +115,8 @@ Full documentation for ROCm Compute Profiler is available at [https://rocm.docs.
 * Fixed L2 read/write/atomic bandwidths on MI350
 * Update metric names for better alignment between analysis configuration and documentation
 * Fixed an issue where accumulation counters could not be collected on AMD Instinct MI100
+* Updated Roofline plots to handle and apply kernel filtering.
+
 
 ### Known issues